Refactor extend parser const generic, refactor run mode
