+++ rust-app/src/interface.rs
@@ -1,9 +1,18 @@
 use core::convert::TryFrom;
+use core::marker::ConstParamTy;
 use ledger_device_sdk::io::{ApduHeader, StatusWords};
 use ledger_parser_combinators::bcs::async_parser::*;
 use ledger_parser_combinators::core_parsers::*;
 use ledger_parser_combinators::endianness::*;
 use num_enum::TryFromPrimitive;
+
+#[derive(ConstParamTy, PartialEq, Eq)]
+#[repr(u8)]
+pub enum ParseChecks {
+    None,
+    PromptUser,
+    CheckSwapTx,
+}
 
 // Payload for a public key request
 pub type Bip32Key = DArray<Byte, U32<{ Endianness::Little }>, 10>;
