+++ rust-app/src/lib.rs
@@ -63,6 +63,9 @@ pub mod main_stax;
 #[cfg(target_family = "bolos")]
 pub mod swap;
 
+#[cfg(target_family = "bolos")]
+pub mod run_mode;
+
 #[cfg(all(target_family = "bolos", test))]
 use core::panic::PanicInfo;
 /// In case of runtime problems, return an internal error and exit the app
