+++ rust-app/src/main_nanos.rs
@@ -1,6 +1,7 @@
 use crate::handle_apdu::*;
 use crate::interface::*;
 use crate::menu::*;
+use crate::run_mode::RunMode;
 use crate::settings::*;
 use crate::ui::UserInterface;
 use crate::swap;
@@ -8,7 +9,6 @@ use crate::swap::get_params::get_check_address_params;
 use crate::swap::get_params::get_printable_amount_params;
 use crate::swap::get_params::sign_tx_params;
 use crate::swap::get_params::swap_return;
-use crate::swap::get_params::CreateTxParams;
 use crate::swap::get_params::SwapResult;
 
 use alamgu_async_block::*;
@@ -23,91 +23,6 @@ use ledger_prompts_ui::{handle_menu_button_event, show_menu};
 use core::cell::RefCell;
 use core::pin::Pin;
 use pin_cell::*;
-
-#[repr(u8)]
-pub enum RunMode {
-    App = 0x00,
-    LibSwapSign {
-        in_progress: bool,
-        is_success: bool,
-        tx_params: CreateTxParams,
-    },
-}
-
-impl RunMode {
-    pub fn is_swap_signing(&self) -> bool {
-        matches!(self, RunMode::LibSwapSign { .. })
-    }
-
-    pub fn start_swap_signing(&mut self, tx_params: CreateTxParams) {
-        debug_assert!(matches!(self, RunMode::App));
-
-        *self = RunMode::LibSwapSign {
-            in_progress: true,
-            is_success: false,
-            tx_params,
-        };
-    }
-
-    pub fn set_signing_result(&mut self, success: bool) {
-        let Self::LibSwapSign {
-            in_progress,
-            is_success,
-            ..
-        } = self
-        else {
-            // Don't care about signing result if we are in `App`` mode
-            return;
-        };
-
-        assert!(*in_progress, "Signing result set when not in progress");
-
-        *in_progress = false;
-        *is_success = success;
-    }
-
-    pub fn is_swap_signing_done(&self) -> bool {
-        matches!(
-            self,
-            RunMode::LibSwapSign {
-                in_progress: false,
-                ..
-            }
-        )
-    }
-
-    pub fn swap_sing_result(&self) -> (bool, *mut u8) {
-        let Self::LibSwapSign {
-            in_progress: false,
-            is_success,
-            tx_params,
-        } = self
-        else {
-            panic!("Not in signing mode or still in progress");
-        };
-
-        (*is_success, tx_params.exit_code_ptr)
-    }
-
-    pub fn tx_params(&self) -> &CreateTxParams {
-        let Self::LibSwapSign { tx_params, .. } = self else {
-            panic!("Not in signing mode");
-        };
-
-        tx_params
-    }
-}
-
-pub struct RunModeInstance;
-
-impl RunModeInstance {
-    pub fn get(&mut self) -> &mut RunMode {
-        static mut RUN_MODE: RunMode = RunMode::App;
-
-        // NOTE: returned lifetime is bound to self and not the 'static
-        unsafe { &mut RUN_MODE }
-    }
-}
 
 #[allow(dead_code)]
 pub fn app_main() {
@@ -152,7 +67,7 @@ pub fn app_main() {
     let menu = |states: core::cell::Ref<'_, Option<APDUsFuture>>,
                 idle: &IdleMenuWithSettings,
                 busy: &BusyMenu| {
-        if RunModeInstance.get().is_swap_signing() {
+        if RunMode.is_swap_signing() {
             return;
         }
 
@@ -165,7 +80,7 @@ pub fn app_main() {
     // Draw some 'welcome' screen
     menu(states.borrow(), &idle_menu, &busy_menu);
     loop {
-        if RunModeInstance.get().is_swap_signing_done() {
+        if RunMode.is_swap_signing_done() {
             comm.borrow_mut().swap_reply_ok();
             return;
         }
@@ -289,9 +204,9 @@ pub fn lib_main(arg0: u32) {
             trace!("{:X?}", params);
             trace!("amount {}", params.amount);
 
-            RunModeInstance.get().start_swap_signing(params);
+            RunMode.start_swap_signing(params);
             app_main();
-            let (is_ok, _) = RunModeInstance.get().swap_sing_result();
+            let (is_ok, _) = RunMode.swap_sing_result();
 
             swap_return(SwapResult::CreateTxResult(&mut params, is_ok as u8));
         }
